Split `ValueBuilder` and family to their own file
