+++ lib/models.dart
@@ -37,6 +37,7 @@ export "src/models/view/modes/arm.dart";
 export "src/models/view/modes/autonomy.dart";
 export "src/models/view/modes/drive.dart";
 export "src/models/view/modes/science.dart";
+export "src/models/view/builder.dart";
 export "src/models/view/settings.dart";
 
 /// A wrapper model around all other data models used by the app.
